Mount every story in the smoke test and document the styling gates
