+++ .github/workflows/ci.yml
@@ -35,5 +35,5 @@ jobs:
       - run: pnpm test
 
       # A story that fails to compile is a tab nobody can look at. Rendering is
-      # covered by src/OscSheet/stories/tabs.smoke.test.tsx, above.
+      # covered by src/OscSheet/stories/stories.smoke.test.tsx, above.
       - run: pnpm storybook:build
